+++ modulo3/projeto-astromatch/src/components/matchPage/MatchPage.js
@@ -1,23 +1,6 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
 import axios from "axios";
-
-const Section = styled.div`
-  display: flex;
-  flex-direction: column;
-  div {
-    display: flex;
-    padding: 10px;
-    img {
-      height: 100%;
-      width: 50px;
-      margin-right: 10px;
-      border-radius: 50%;
-      cursor: pointer;
-      
-    }
-  }
-`;
+import { Section } from "./styled.Match";
 
 const MatchPage = (props) => {
   const [matcheList, setMacheList] = useState([]);
@@ -42,17 +25,19 @@ const MatchPage = (props) => {
     return (
       <Section key={matche.id}>
         <div>
-          <img src={matche.photo}/>
+          <img src={matche.photo} />
           <p>{matche.name}</p>
         </div>
       </Section>
     );
   });
-  
-  return <div>
-    <h2>Matches</h2>
-    {matchesList}
-    </div>;
+
+  return (
+    <div>
+      <h2>Matches</h2>
+      {matchesList}
+    </div>
+  );
 };
 
 export default MatchPage;
