+++ modulo3/projeto-astromatch/src/components/profilePage/ProfilesPage.js
@@ -1,55 +1,8 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
 import axios from "axios";
-import ScreenReset from "./ScreenReset";
-
-export const Section = styled.section`
-  padding: 20px 20px 0px;
-  display: flex;
-  flex-direction: column;
-  flex: 1 1 0%;
-  -webkit-box-pack: end;
-  justify-content: flex-end;
-`;
-const DivMatch = styled.div`
-  box-shadow: rgb(117 117 117 / 77%) 0px 2px 10px 0px;
-  position: relative;
-  border-radius: 5px;
-  overflow: hidden;
-  transition: all 0.5s ease 0s;
-  height: 430px;
-  animation: 0.5s ease 0s 1 normal forwards running none;
-  display: flex;
-  -webkit-box-align: center;
-  align-items: center;
-  img {
-    width: 100%;
-    height: 100%;
-    /* display: block; */
-    z-index: 1;
-  }
-`;
-const Selectors = styled.div`
-  display: flex;
-  justify-content: space-evenly;
-  -webkit-box-align: center;
-  align-items: center;
-  padding: 10px 0px;
-`;
-const Texto = styled.div`
-  height: 30%;
-  position: absolute;
-  bottom: 0;
-  width: 92%;
-  background-image: linear-gradient(to top, rgba(0, 0, 0, 0.5), transparent);
-  color: white;
-  display: flex;
-  flex-direction: column;
-  -webkit-box-pack: end;
-  justify-content: flex-end;
-  padding: 15px;
-  z-index: 2;
-`;
+import ScreenReset from "../screenReset/ScreenReset";
+import {Section, DivMatch, Selectors, Texto, BotonLik} from "./styledProfilePage"
+import {GrAed} from 'react-icons/gr'
 
 const ProfilesPage = (props) => {
   const [profile, setProfile] = useState({});
@@ -88,7 +41,7 @@ const ProfilesPage = (props) => {
         response.data.isMatch === true && alert("Deu Match!");
       })
       .catch((err) => {
-        // console.log(err);
+        console.log(err);
       });
   };
 
@@ -125,12 +78,13 @@ const ProfilesPage = (props) => {
           </DivMatch>
           <Selectors>
             <button onClick={getProfile}>Dislike</button>
-            <button onClick={postProfile}>Like</button>
+            <BotonLik onClick={postProfile}><GrAed/></BotonLik>
           </Selectors>
         </Section>
       ) : (
         <ScreenReset resetProfile={resetProfile} />
       )}
+      {/* <ScreenReset resetProfile={resetProfile} /> */}
     </div>
   );
 };
